Fix signed vs. unsigned warnings.
Most of the warnings in GaussianProcess.cc were patched with
on-the-spot casts. While changing the internal members to unsigned
makes more sense a priori, the current code uses -1 as a null value,
and it is very difficult to maintain KdTree's class invariants during
the conversion. It may be less work to just replace the implementation
of KdTree with an object-oriented rather than array-oriented
implementation.

+++ src/math/Interpolate.cc
@@ -301,7 +301,7 @@ std::vector<double> Interpolate::interpolate(std::vector<double> const &x) const
 ndarray::Array<double, 1> Interpolate::interpolate(ndarray::Array<double const, 1> const &x) const {
     int const num = x.getShape()[0];
     ndarray::Array<double, 1> out = ndarray::allocate(ndarray::makeVector(num));
-    for (size_t i = 0; i < num; ++i) {
+    for (int i = 0; i < num; ++i) {
         std::cout << "Interpolating " << x[i] << std::endl;
         out[i] = interpolate(x[i]);
     }
